+++ app/src/main/kotlin/me/jerryokafor/ihenkiri/ui/MainActivity.kt
@@ -141,7 +141,8 @@ class MainActivity : ComponentActivity() {
             Log.d(TAG, "ProfileInstaller status code: ${status.profileInstallResultCode}")
             val compilationStatus = when {
                 status.isCompiledWithProfile -> "ProfileInstaller: is compiled with profile"
-                status.hasProfileEnqueuedForCompilation() -> "ProfileInstaller: Enqueued for compilation"
+                status.hasProfileEnqueuedForCompilation() ->
+                    "ProfileInstaller: Enqueued for compilation"
                 else -> "Profile not compiled or enqueued"
             }
             Log.d(TAG, compilationStatus)
