+++ feature/peopleDetails/src/main/java/com/jerryokafor/feature/peopledetails/ui/timeline/TimelineNode.kt
@@ -22,20 +22,38 @@
  * THE SOFTWARE.
  */
 
-package com.jerryokafor.feature.peopledetails.ui
+@file:Suppress("MagicNumber")
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2023 IheNkiri Project
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.Arrangement
+package com.jerryokafor.feature.peopledetails.ui.timeline
+
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithCache
 import androidx.compose.ui.geometry.Offset
@@ -46,93 +64,10 @@ import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.drawscope.clipRect
 import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.drawText
 import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import kotlinx.coroutines.launch
-import me.jerryokafor.core.ds.theme.IheNkiri
-
-data class TimelineNodeConfig(
-    val indicatorColor: Color,
-    val circleRadius: Dp,
-    val circleStrokeWidth: Dp,
-    val lineStrokeWidth: Dp,
-    val titleTextStyle: TextStyle,
-)
-
-object TimelineNodeDefaults {
-    private val indicatorColor: Color = Color.White
-    private val circleRadius: Dp = 8.dp
-    private val circleStrokeWidth: Dp = 3.0.dp
-    private val lineStrokeWidth: Dp = 1.0.dp
-    private val titleTextStyle: TextStyle = TextStyle(fontSize = 14.sp)
-
-    @Composable
-    fun timelineConfig(
-        indicatorColor: Color = TimelineNodeDefaults.indicatorColor,
-        circleRadius: Dp = TimelineNodeDefaults.circleRadius,
-        circleStrokeWidth: Dp = TimelineNodeDefaults.circleStrokeWidth,
-        lineStrokeWidth: Dp = TimelineNodeDefaults.lineStrokeWidth,
-        titleTextStyle: TextStyle = TimelineNodeDefaults.titleTextStyle,
-    ): TimelineNodeConfig = TimelineNodeConfig(
-        indicatorColor = indicatorColor,
-        circleRadius = circleRadius,
-        circleStrokeWidth = circleStrokeWidth,
-        lineStrokeWidth = lineStrokeWidth,
-        titleTextStyle = titleTextStyle,
-    )
-}
-
-interface TimelineScope {
-    val title: String
-    val content: @Composable BoxScope.(Modifier) -> Unit
-}
-
-@Composable
-fun <T : TimelineScope> Timeline(
-    modifier: Modifier = Modifier,
-    items: List<T>,
-) {
-    val lineAnim = remember { Animatable(0f) }
-    val circleAnim = remember { Animatable(0f) }
-
-    Column(modifier = modifier, verticalArrangement = Arrangement.Center) {
-        items.forEach { item ->
-            TimelineNode(
-                title = item.title,
-                lineAnim = lineAnim.value,
-                circleAnim = circleAnim.value,
-                config = TimelineNodeDefaults.timelineConfig(
-                    titleTextStyle = IheNkiri.typography.titleMedium,
-                    indicatorColor = IheNkiri.color.primary,
-                ),
-            ) {
-                item.content(this, it)
-            }
-        }
-    }
-
-    LaunchedEffect(Unit, block = {
-        launch {
-            lineAnim.animateTo(
-                targetValue = 1f,
-                animationSpec = @Suppress("MagicNumber")
-                (tween(2000, easing = LinearEasing)),
-            )
-        }
-        launch {
-            circleAnim.animateTo(
-                targetValue = 1f,
-                animationSpec = @Suppress("MagicNumber")
-                (tween(2000, easing = LinearEasing)),
-            )
-        }
-    })
-}
 
 @Composable
 fun TimelineNode(
